build: breaking down GetDirections, adding markers and zoom to direction
